[Feat] 가입한 그룹 개별 조회 및 가입한 그룹원 캘린더 조회

+++ src/main/java/scs/planus/domain/group/dto/GroupCreateRequestDto.java
@@ -22,5 +22,5 @@ public class GroupCreateRequestDto {
     @NotNull(message = "[request] 그룹 제한 인원을 설정해 주세요.")
     @Max(value = 50, message = "[request] 그룹 제한 인원은 50명 이하로 입력해 주세요.")
     @Min(value = 2, message = "[request] 그룹 제한 인원은 2명 이상으로 입력해 주세요.")
-    private Long limitCount;
+    private int limitCount;
 }
